spring rest

+++ practice_spring/src/main/java/com/test/mapper/ReplyMapper.java
@@ -11,7 +11,7 @@ public interface ReplyMapper {
 	
 	public List<ReplyVO> getListWithPaging(@Param("bno")Long bno);
 	
-	public int delete (int rno);
+	public int delete (Long rno);
 	
 	public int modify(ReplyVO vo);
